MessageReceiveTimeout(): change timeOut type from int32_t to unsigned long to avoid overflows with a timeout of 1000000000.
The bug was that MessageReceiveTimeout() timeouts immediatly when it
should not. So a _lot_ of exchanges with pcscd occurs.

Tested on 64-bits CPU only.


git-svn-id: svn://anonscm.debian.org/svn/pcsclite/trunk/PCSC@5796 0ce88b0d-b2fd-0310-8134-9614164e65ea

+++ src/winscard_msg.h
@@ -258,7 +258,7 @@
 	int32_t ClientSetupSession(uint32_t *);
 	int32_t ClientCloseSession(uint32_t);
 	LONG MessageReceiveTimeout(uint32_t command, /*@out@*/ void *buffer,
-		uint64_t buffer_size, int32_t filedes, int32_t timeOut);
+		uint64_t buffer_size, int32_t filedes, unsigned long timeOut);
 	LONG MessageSendWithHeader(uint32_t command, uint32_t dwClientID,
 		uint64_t size, void *data);
 #endif
